+++ Car_System_BackEnd/src/main/java/com/carsystem/app/model/RentalPackage.java
@@ -6,7 +6,6 @@ import lombok.NoArgsConstructor;
 import jakarta.persistence.*;
 
 import java.time.LocalDateTime;
-import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -21,8 +20,9 @@ public class RentalPackage {
     private String name;
     private String description;
     private int durationHours;
-    private BigDecimal basePrice;
+    private Double basePrice;
     private boolean isActive;
+    
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
 }
